stuck page classes in their own files
